連載ナビに出ている記事を関連記事・被参照記事から落とす
記事末尾は 連載ナビ -> 関連記事 -> 被参照記事 の順に並ぶ。ナビが既に
出している記事がすぐ下でもう一度出ると、読者は同じ導線を二度見る。

実測では連載ナビのある620記事のうち、被参照記事と重複610件
（うち索引が538件）、関連記事と重複75件だった。索引は連載の全記事を
参照するので、連載の記事はほぼ必ず被参照に索引が並んでいた。

連載のグループ化を lib/series.js に移し、ナビ本体・関連記事・被参照記事の
3つが同じ結果を見るようにする。ナビがリンクしている索引・前・次の
パスを持たせ、両方の一覧から落とす。

関連記事は620記事すべてが3件を保つ（カテゴリからの補填が効く）。
被参照記事が空になる記事は6件から386件に増える。索引からしか
参照されていない記事で、その索引はナビの見出しから辿れる。

Closes #2054
